requirements.txt
PyYAML==6.0
aiogram==2.19
aioschedule==0.5.2